dposterior

Compute (non-normalized) posterior densities for a given parameter set.

dposterior(posterior_sample, data, theta = NULL)

Arguments

posterior_sample
a list given by the Bliss_Gibbs_Sampler function.

data
a list containing

y

a numerical vector, the outcomes.

x

a list of matrices, the qth matrix contains the observations of the qth functional covariate at time points given by grids.

theta
a matrix or a vector which contains the parameter set.

Value

Return the (log) posterior density, the (log) likelihood and the (log) prior density for the given parameter set.

Details

If the theta is NULL, the posterior density is computed from the MCMC sample given in the posterior_sample.

Examples

data(data1)
data(param1)
# result of res_bliss1<-fit_Bliss(data=data1,param=param1)
data(res_bliss1)
# Compute the posterior density of the MCMC sample :
res_poste <- dposterior(res_bliss1$posterior_sample,data1)