predict_bliss

Compute predictions.

predict_bliss(x, grids, burnin, posterior_sample, Smooth_estimate)

Arguments

x

a list containing the design matrices related to the functional covariates. Must be similar to the result of the function sim_x.

grids

a list of numerical vectors, the qth vector is the grid of time points for the qth functional covariate.

burnin

an integer (optional), the number of iteration to drop from the posterior sample.

posterior_sample

a list provided by the function Bliss_Gibbs_Sampler.

Smooth_estimate

one of the objects resulting from Bliss_Simulated_Annealing.

Value

A vector of predictions for each individual data x.

Examples

# \donttest{
data(data1)
data(param1)
data(res_bliss1)

predict_bliss(data1$x,data1$grids,50,res_bliss1$posterior_sample,res_bliss1$smooth_estimate)
#>   [1]  0.89867036  0.86370969  0.62740505  2.55012603  0.95599063  1.46801839
#>   [7]  2.11455626  1.94417819  0.48805741  0.71640667  0.30513392  1.31826834
#>  [13] -0.10724065  1.00168385 -0.31996668 -0.45586517  0.22148405 -0.03516862
#>  [19]  1.80923513  1.88033033  1.91771163  0.62279761 -0.39364420  0.33165896
#>  [25]  2.33750899  1.26785422  0.57444461  1.39591982  0.62803506  0.73553607
#>  [31]  1.65303991  1.16943611  0.37726843  1.75881548  0.76487590  2.18851967
#>  [37]  2.15626983  0.48058949  0.17438501  0.32824148  0.16148622  1.58712171
#>  [43]  1.86741131  0.96800406  0.50845539  0.72523935  2.29398900  2.14644410
#>  [49]  1.54181905  0.17936007  1.40143768  1.59685988  2.54376060  1.74910729
#>  [55]  1.43143900 -0.44821959  1.56941628  0.35751017  1.03051738  1.62594643
#>  [61]  1.74894841  1.59494358 -0.22883328  1.58059554  2.20660529 -0.02313381
#>  [67] -0.15681712  0.62878783 -0.14909036  1.07811602 -0.41683076  1.13464409
#>  [73]  1.61499743  0.35896604 -0.74197240  0.75140358  0.73122894  1.86063781
#>  [79]  1.42455055 -0.02706995  0.54923276  0.88470642  1.21019282 -0.13539527
#>  [85]  1.16525408  1.48750599  1.75170888  0.69124918  0.64027150  0.79046607
#>  [91]  1.74248139  0.01802965  2.08190166  2.00838463  1.93219044  1.60120357
#>  [97] -0.06904372 -0.72671674 -0.04765908  0.35059550
# }