A result of the BliSS method

A result of the BliSS method

res_bliss1

Format

a Bliss object (list)

alpha

a list of Q numerical vector. Each vector is the function alpha(t) associated to a functional covariate. For each t, alpha(t) is the posterior probabilities of the event "the support covers t".

beta_posterior_density

a list of Q items. Each item contains a list containing information to plot the posterior density of the coefficient function with the image function.

grid_t

a numerical vector: the x-axis.

grid_beta_t

a numerical vector: the y-axis.

density

a matrix: the z values.

new_beta_sample

a matrix: beta sample used to compute the posterior densities.

beta_sample

a list of Q matrices. The qth matrix is a posterior sample of the qth functional covariates.

Bliss_estimate

a list of numerical vectors corresponding to the Bliss estimates of each functional covariates.

chains_info

a list containing (for each chain): a mu estimate, a sigma_sq estimate, the Smooth estimate of the coefficient function and the autocorrelation of the Markov Chain.

data

see the description of the object data1.

posterior_sample

a list containing (for each chain) the result of the Bliss_Gibbs_Sampler function.

Smooth_estimate

a list containing the Smooth estimates of the coefficient functions.

support_estimate

a list containing the estimations of the support.

support_estimate_fct

a list containing the estimation of the support.

trace_sann

a list containing (for each chain) the trace of the Simulated Annealing algorithm.